+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Added better in-line documentation.
 - Added new string option - `type` - to `leaderboard` command, can now either choose between `Sent` or `Received`
   emojis.
+- Added new command `/resetdb`
 
 ### Changed
 
@@ -28,6 +29,7 @@
 - `messageUpdate.js` Thrown errors properly caught.
 - `listemojis.js` Thrown errors properly caught.
 - `leaderboard.js` Fixed hourly daterange.
+- `displaystats.js` Fixed empty output if no records in databases.
 
 ## v1.4.3 - 2021-09-28